Split out ESPHSVColor into a separate file
